Add VulnManagementCardList for responsive vulnerability display
